blogapi/users/serializers.py
from rest_framework import serializers
from .models import User

class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'prenom', 'nom', 'login', 'email', 'created_at']